@@ -4,6 +4,7 @@
   "description": "Create Stylish JavaScript web app",
   "author": "StyleList94",
   "license": "MIT",
+  "type": "module",
   "repository": {
     "type": "git",
     "url": "https://github.com/StyleList94/create-stylish-app.git"
@@ -12,11 +13,11 @@
     "url": "https://github.com/StyleList94/create-stylish-app/issues"
   },
   "engines": {
-    "node": ">=18"
+    "node": ">=18.18.0"
   },
   "dependencies": {
-    "chalk": "^4.1.2",
-    "commander": "^9.5.0",
+    "chalk": "^5.3.0",
+    "commander": "^12.1.0",
     "cross-spawn": "^7.0.3"
   },
   "files": [
