@@ -2,37 +2,52 @@
 
 'use strict';
 
-const path = require('path');
-const fs = require('fs');
-const https = require('https');
-const { Command } = require('commander');
-const chalk = require('chalk');
-const spawn = require('cross-spawn');
-const packageJson = require('./package.json');
+import { resolve as _resolve, join } from 'node:path';
+import { readFileSync, writeFileSync, mkdirSync } from 'node:fs';
+import { get } from 'node:https';
+import { Command } from 'commander';
+import chalk from 'chalk';
+import { sync } from 'cross-spawn';
+
+function getPackageManager() {
+  const userAgent = process.env.npm_config_user_agent || '';
+
+  if (userAgent.startsWith('yarn')) {
+    return 'yarn';
+  }
+
+  if (userAgent.startsWith('pnpm')) {
+    return 'pnpm';
+  }
+
+  if (userAgent.startsWith('bun')) {
+    return 'bun';
+  }
+
+  return 'npm';
+}
 
 function checkForValidTemplate(template) {
   return new Promise((resolve, reject) => {
-    https
-      .get(
-        `https://github.com/StyleList94/stylish-${template}-${
-          template === 'ethereum' ? 'd' : ''
-        }app`,
-        (res) => {
-          if (res.statusCode === 200) {
-            resolve(true);
-          } else {
-            resolve(false);
-          }
+    get(
+      `https://github.com/StyleList94/stylish-${template}-${
+        template === 'ethereum' ? 'd' : ''
+      }app`,
+      (res) => {
+        if (res.statusCode === 200) {
+          resolve(true);
+        } else {
+          resolve(false);
         }
-      )
-      .on('error', (e) => {
-        reject(e);
-      });
+      }
+    ).on('error', (e) => {
+      reject(e);
+    });
   });
 }
 
 function execCommand(command, args, options = { silent: false }) {
-  const result = spawn.sync(command, args, {
+  const result = sync(command, args, {
     stdio: options.silent ? 'ignore' : 'inherit',
   });
 
@@ -44,7 +59,7 @@ function execCommand(command, args, options = { silent: false }) {
 
 function buildPackageJson(packageJsonPath, appName) {
   try {
-    const fileContent = fs.readFileSync(packageJsonPath, 'utf-8');
+    const fileContent = readFileSync(packageJsonPath, 'utf-8');
     const projectPackageJson = JSON.parse(fileContent);
 
     const { description, author, ...nextPackageJson } = projectPackageJson;
@@ -54,8 +69,8 @@ function buildPackageJson(packageJsonPath, appName) {
       version: '0.1.0',
     });
 
-    fs.writeFileSync(
-      path.resolve(process.cwd(), 'package.json'),
+    writeFileSync(
+      _resolve(process.cwd(), 'package.json'),
       JSON.stringify(nextPackageJson, null, 2),
       'utf8'
     );
@@ -66,6 +81,8 @@ function buildPackageJson(packageJsonPath, appName) {
 }
 
 function getStartScript(template) {
+  const packageManager = getPackageManager();
+
   switch (template) {
     case 'react':
     case 'next':
@@ -73,21 +90,21 @@ function getStartScript(template) {
     case 'ethereum':
     case 'web':
     case 'pure-react':
-      return 'yarn dev';
+      return `${packageManager} run dev`;
 
     default:
       return '';
   }
 }
 
-async function run(appName, template) {
+async function run(appName, packageInfo, template) {
   try {
     const isValidTemplate = await checkForValidTemplate(template);
     if (!isValidTemplate) {
-      console.error(`Template ${template} is not valid`);
+      console.error(`Template ${chalk.bold(template)} is not valid`);
       console.log();
       console.log(
-        `Run ${chalk.cyan(`${packageJson.name} --help`)} to see all options.`
+        `Run ${chalk.cyan(`${packageInfo.name} --help`)} to see all options.`
       );
       process.exit(1);
     }
@@ -97,35 +114,43 @@ async function run(appName, template) {
   }
 
   const cwd = process.cwd();
-  const appPath = path.join(cwd, appName);
+  const appPath = join(cwd, appName);
   const repository = `https://github.com/StyleList94/stylish-${template}-${
     template === 'ethereum' ? 'd' : ''
   }app.git`;
 
-  console.log(`Creating a new ${chalk.cyan(`stylish-${template}-app`)}`);
+  console.log(`\nCreating a new ${chalk.cyan(`stylish-${template}-app`)}`);
   console.log(`in ${chalk.green(appPath)}\n`);
   try {
-    fs.mkdirSync(appPath);
+    mkdirSync(appPath);
   } catch (error) {
     if (error.code === 'EEXIST') {
       console.log(
-        `Error\nThe app name ${appName} is already exist in the current directory, please change the app name.\n`
+        `${chalk.bgRed('Error!')}\nThe app name ${chalk.bold(
+          appName
+        )} is already exist in the current directory, please change app name.\n`
       );
     } else {
       console.log(error);
     }
     process.exit(1);
   }
-  execCommand('git', ['clone', repository, appName], { silent: true });
+  process.stdout.write(`\nClone repository...`);
+  execCommand('git', ['clone', '--depth=1', repository, appName], {
+    silent: true,
+  });
   process.chdir(appPath);
-  buildPackageJson(path.resolve(process.cwd(), 'package.json'), appName);
+  buildPackageJson(_resolve(process.cwd(), 'package.json'), appName);
+  console.log(` Done!\n`);
 
-  if (template !== 'react') {
-    console.log(
-      `\nInstalling packages. This might take a couple of minutes...\n`
-    );
-    execCommand('yarn', ['install']);
+  const packageManager = getPackageManager();
+
+  if (packageManager !== 'yarn') {
+    execCommand('npx', ['rimraf', './yarn.lock'], { silent: true });
   }
+
+  console.log(`Run ${chalk.cyan.italic(`${packageManager} install`)}...\n`);
+  execCommand(packageManager, ['install']);
 
   console.log(`\nInitializing git...\n`);
   execCommand('npx', ['rimraf', './.git'], { silent: true });
@@ -134,12 +159,9 @@ async function run(appName, template) {
   execCommand('git', ['commit', '-m', 'initial commit'], { silent: true });
   execCommand('git', ['branch', '-m', 'main'], { silent: true });
 
-  console.log(`All done.\n`);
+  console.log(`\nAll Done!\n`);
   console.log('Run the following commands to get started:\n');
   console.log(`  ${chalk.cyan('cd')} ${appName}`);
-  if (template === 'react') {
-    console.log(`  ${chalk.cyan('yarn')} install`);
-  }
   console.log(`  ${chalk.cyan(getStartScript(template))}`);
   console.log();
 }
@@ -147,11 +169,15 @@ async function run(appName, template) {
 function init() {
   let appName;
 
-  const program = new Command(packageJson.name);
+  const packageInfo = JSON.parse(
+    readFileSync(new URL('./package.json', import.meta.url), 'utf8')
+  );
+
+  const program = new Command(packageInfo.name);
 
   program
     .description(chalk.cyan('Create Stylish JavaScript web app'))
-    .version(packageJson.version, '-v, --version')
+    .version(packageInfo.version, '-v, --version')
     .arguments('[app-name]')
     .usage('<app-name> [options]')
     .option(
@@ -182,7 +208,7 @@ function init() {
 
   const options = program.opts();
 
-  run(appName, options.template);
+  run(appName, packageInfo, options.template);
 }
 
 init();
